# src/main/webapp/api/openapi.json
{
  "components" : {
    "schemas" : {
      "BoardView" : {
        "properties" : {
          "tasks" : {
            "items" : {
              "$ref" : "#/components/schemas/TaskView"
            },
            "type" : "array"
          },
          "zones" : {
            "items" : {
              "$ref" : "#/components/schemas/ZoneLoad"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "CaptureTask" : {
        "properties" : {
          "clientRef" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          },
          "dueDate" : {
            "format" : "date",
            "type" : "string"
          },
          "labels" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "notes" : {
            "maxLength" : 10000,
            "minLength" : 0,
            "type" : "string"
          },
          "title" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "zone" : {
            "enum" : [ "CRITICAL_NOW", "OPPORTUNITY_NOW", "OVER_THE_HORIZON" ],
            "type" : "string"
          }
        },
        "required" : [ "title" ],
        "type" : "object"
      },
      "CreateTask" : {
        "properties" : {
          "clientRef" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          },
          "dueDate" : {
            "format" : "date",
            "type" : "string"
          },
          "labels" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "notes" : {
            "maxLength" : 10000,
            "minLength" : 0,
            "type" : "string"
          },
          "title" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "zone" : {
            "enum" : [ "CRITICAL_NOW", "OPPORTUNITY_NOW", "OVER_THE_HORIZON" ],
            "type" : "string"
          }
        },
        "required" : [ "title", "zone" ],
        "type" : "object"
      },
      "CreateTaskList" : {
        "properties" : {
          "name" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
      "Defer" : {
        "properties" : {
          "until" : {
            "format" : "date",
            "type" : "string"
          }
        },
        "required" : [ "until" ],
        "type" : "object"
      },
      "MoveZone" : {
        "properties" : {
          "zone" : {
            "enum" : [ "CRITICAL_NOW", "OPPORTUNITY_NOW", "OVER_THE_HORIZON" ],
            "type" : "string"
          }
        },
        "required" : [ "zone" ],
        "type" : "object"
      },
      "RenameTaskList" : {
        "properties" : {
          "name" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
      "SetLabels" : {
        "properties" : {
          "labels" : {
            "items" : {
              "maxLength" : 64,
              "minLength" : 0,
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "labels" ],
        "type" : "object"
      },
      "Share" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "email" ],
        "type" : "object"
      },
      "TaskListDetail" : {
        "properties" : {
          "list" : {
            "$ref" : "#/components/schemas/TaskListSummary"
          },
          "tasks" : {
            "items" : {
              "$ref" : "#/components/schemas/TaskView"
            },
            "type" : "array"
          },
          "zones" : {
            "items" : {
              "$ref" : "#/components/schemas/ZoneLoad"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "TaskListSummary" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "inbox" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          },
          "owned" : {
            "type" : "boolean"
          },
          "sharedWith" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "slug" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "TaskView" : {
        "properties" : {
          "clientRef" : {
            "type" : "string"
          },
          "deferUntil" : {
            "format" : "date",
            "type" : "string"
          },
          "dueDate" : {
            "format" : "date",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "labels" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "lastReviewedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "listId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "listName" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "state" : {
            "enum" : [ "TODO", "DONE" ],
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "zone" : {
            "enum" : [ "CRITICAL_NOW", "OPPORTUNITY_NOW", "OVER_THE_HORIZON" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "UpdateTask" : {
        "properties" : {
          "dueDate" : {
            "format" : "date",
            "type" : "string"
          },
          "notes" : {
            "maxLength" : 10000,
            "minLength" : 0,
            "type" : "string"
          },
          "title" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ZoneLoad" : {
        "properties" : {
          "open" : {
            "format" : "int64",
            "type" : "integer"
          },
          "overSoftCap" : {
            "type" : "boolean"
          },
          "softCap" : {
            "format" : "int32",
            "type" : "integer"
          },
          "zone" : {
            "enum" : [ "CRITICAL_NOW", "OPPORTUNITY_NOW", "OVER_THE_HORIZON" ],
            "type" : "string"
          }
        },
        "type" : "object"
      }
    }
  },
  "info" : {
    "title" : "One Minute To-Do List",
    "version" : "1"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/board" : {
      "get" : {
        "operationId" : "board",
        "parameters" : [ {
          "in" : "query",
          "name" : "list",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "label",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "zone",
          "required" : false,
          "schema" : {
            "enum" : [ "CRITICAL_NOW", "OPPORTUNITY_NOW", "OVER_THE_HORIZON" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "includeDone",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BoardView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
    "/api/labels" : {
      "get" : {
        "operationId" : "allLabels",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "type" : "string"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
    "/api/review" : {
      "get" : {
        "operationId" : "review",
        "parameters" : [ {
          "in" : "query",
          "name" : "list",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "label",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TaskView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
    "/api/tasklists" : {
      "get" : {
        "operationId" : "allLists",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TaskListSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-list-controller" ]
      },
      "post" : {
        "operationId" : "createList",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTaskList"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskListSummary"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "task-list-controller" ]
      }
    },
    "/api/tasklists/by-slug/{slug}" : {
      "get" : {
        "operationId" : "listBySlug",
        "parameters" : [ {
          "in" : "path",
          "name" : "slug",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskListSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-list-controller" ]
      }
    },
    "/api/tasklists/{id}" : {
      "delete" : {
        "operationId" : "deleteList",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "task-list-controller" ]
      },
      "get" : {
        "operationId" : "oneList",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskListDetail"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-list-controller" ]
      },
      "patch" : {
        "operationId" : "renameList",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RenameTaskList"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskListSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-list-controller" ]
      }
    },
    "/api/tasklists/{id}/shares" : {
      "delete" : {
        "operationId" : "unshareList",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Share"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskListSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-list-controller" ]
      },
      "post" : {
        "operationId" : "shareList",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Share"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskListSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-list-controller" ]
      }
    },
    "/api/tasklists/{listId}/review" : {
      "get" : {
        "operationId" : "reviewInList",
        "parameters" : [ {
          "in" : "path",
          "name" : "listId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TaskView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
    "/api/tasklists/{listId}/tasks" : {
      "get" : {
        "operationId" : "tasksInList",
        "parameters" : [ {
          "in" : "path",
          "name" : "listId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TaskView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      },
      "post" : {
        "operationId" : "addTaskToList",
        "parameters" : [ {
          "in" : "path",
          "name" : "listId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTask"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
    "/api/tasks/capture" : {
      "post" : {
        "operationId" : "captureTask",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CaptureTask"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
    "/api/tasks/{id}" : {
      "delete" : {
        "operationId" : "deleteTask",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "task-controller" ]
      },
      "get" : {
        "operationId" : "oneTask",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      },
      "patch" : {
        "operationId" : "updateTask",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTask"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
    "/api/tasks/{id}/complete" : {
      "post" : {
        "operationId" : "completeTask",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
    "/api/tasks/{id}/defer" : {
      "post" : {
        "operationId" : "deferTask",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Defer"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
    "/api/tasks/{id}/labels" : {
      "put" : {
        "operationId" : "setTaskLabels",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SetLabels"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
    "/api/tasks/{id}/reopen" : {
      "post" : {
        "operationId" : "reopenTask",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
    "/api/tasks/{id}/reviewed" : {
      "post" : {
        "operationId" : "markTaskReviewed",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      }
    },
    "/api/tasks/{id}/zone" : {
      "post" : {
        "operationId" : "moveTaskZone",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MoveZone"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "task-controller" ]
      }
    }
  }
}